requirements.txt
aiodns==2.0.0
aiogram==2.9.2
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
idna==3.1
multidict==5.1.0
pycares==3.1.1
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
typing-extensions==3.7.4.3
ujson==4.0.1
uvloop==0.14.0
yarl==1.6.3
